The death of TAggregable
The abomination can finally be laid to rest.

Also got rid of the AST.scala file.

Also got rid of EvalContext.
