+++ src/main/scala/is/hail/utils/DoubleNumericConversion.scala
@@ -1,0 +1,10 @@
+package is.hail.utils
+
+object DoubleNumericConversion {
+  def to(numeric: Any): Double = numeric match {
+    case i: Int => i
+    case l: Long => l
+    case f: Float => f
+    case d: Double => d
+  }
+}
