+++ src/main/scala/is/hail/utils/Graph.scala
@@ -3,7 +3,7 @@ package is.hail.utils
 import is.hail.annotations.{Region, RegionValueBuilder, SafeRow}
 import is.hail.expr.ir.{Compile, MakeTuple}
 import is.hail.expr.types._
-import is.hail.expr.{EvalContext, IRParserEnvironment, Parser}
+import is.hail.expr.{IRParserEnvironment, Parser}
 import org.apache.spark.sql.Row
 
 import scala.collection.mutable
